security add on HM.

+++ airBndApp/src/main/java/com/example/airBndApp/Service/Impl/RoomServiceimpl.java
@@ -3,15 +3,18 @@ package com.example.airBndApp.Service.Impl;
 import com.example.airBndApp.Dto.RoomDto;
 import com.example.airBndApp.Entity.HotelEntity;
 import com.example.airBndApp.Entity.RoomEntity;
+import com.example.airBndApp.Entity.UserEntity;
 import com.example.airBndApp.Repository.HotelRepo;
 import com.example.airBndApp.Repository.RoomRepo;
 import com.example.airBndApp.Service.InventoryService;
 import com.example.airBndApp.Service.RoomService;
 import com.example.airBndApp.exception.ResourcesNotFoundException;
+import com.example.airBndApp.exception.UnAuthorisedException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -39,6 +42,11 @@ public class RoomServiceimpl implements RoomService {
     public RoomDto createNewRoom(Long hotelId,RoomDto roomDto) {//create a room with details
         log.info("start find hotel");
         HotelEntity hotel = hotelRepo.findById(hotelId).orElseThrow(() -> new ResourcesNotFoundException("Hotel id not Found : "+ hotelId));
+        UserEntity user =(UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())){
+            throw  new UnAuthorisedException("This user does not own this room. user id : "+ user.getId());
+        }
 
         log.info("Room DTO: " + roomDto);
         RoomEntity roomEntity = modelMapper.map(roomDto,RoomEntity.class);
@@ -76,6 +84,13 @@ public class RoomServiceimpl implements RoomService {
     @Transactional
     public void deleteRoomById(Long id) { // delete hotel using id
         RoomEntity room = roomRepo.findById(id).orElseThrow(() -> new ResourcesNotFoundException("room id not Found : "+ id));
+
+        UserEntity user =(UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(room.getHotel().getOwner())){
+            throw  new UnAuthorisedException("This user does not own this room. user id : "+ user.getId());
+        }
+
         log.info("inventory delete : start");
         inventoryService.deleteAllInventory(room);
         log.info("room deleted now ");
